requirements.txt
aspy.yaml==1.3.0
backports.functools-lru-cache==1.5
cachetools==3.1.1
certifi==2019.3.9
cfgv==2.0.0
chardet==3.0.4
cloudpickle==1.1.1
configparser==3.7.4
contextlib2==0.5.5
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
enum34==1.1.6
filelock==3.0.12
futures==3.2.0
grpcio==1.21.1
identify==1.4.3
idna==2.8
image==1.5.27
imageio==2.4.1
importlib-metadata==0.17
importlib-resources==1.0.2
kiwisolver==1.1.0
lxml==4.3.3
matplotlib==2.2.4
moviepy==1.0.0
networkx==2.2
nodeenv==1.3.3
numpy==1.16.3
opencv-python==4.1.0.25
pathlib2==2.3.3
pluggy==0.12.0
pre-commit==1.16.1
proglog==0.1.9
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
pysrt==1.1.1
python-dateutil==2.8.0
python-vlc==3.0.6109
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
requests==2.22.0
rsa==4.0
scandir==1.10.0
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
subprocess32==3.5.4
toml==0.10.0
toolz==0.9.0
tox==3.12.1
tqdm==4.32.1
typing==3.6.6
urllib3==1.25.3
virtualenv==16.6.0
